Close #1: Remove instance observers upon de-allocation
The only way for Class::Observable to know when an instance is
de-allocated, is by hooking into DESTROY.  It's better to entirely
delete the key from the hash as the chances of the entry being re-used
(i.e. an instance of the same type being allocated on exactly the same
address) is low; leaving an empty arrayref does not make sense in that
respect.